result: replace r type with result type
The r type had two major drawbacks. First, the use of polymorphic
variants made type errors hard to understand. Second, the names of the
polymorphic variants themselves were somewhat confusing, in particular
Cont.

The r type is now replaces with the result type, which is equivalent to
the Result.t type from Core, and other libraires. This will be the type
that the community will standardize on, so use that for now. See
ocaml/ocaml#147.