auth: Support typical auth schemes for is_authorized return type
The Erlang implementation only allows a resource to indicate that the
request is authorized, or to issue a authentication challenge for the
Basic authorization scheme, given a realm. This implementation supports
those two modes, as well as an additional one. A resource may specify a
URI that the user-agent should visit where further authentication
instructions will be present.