accept: Don't crash on parsing errors
Cohttp.Accept.media_ranges raises exceptions when it can't parse the
`Accept` header. Notably, whenever the format isn't: (token)+/(token)+,
it raises an exception. Since we weren't capturing it, it went on to
crash the server.

This patch makes it treat any parsing exception in the `Accept` header
as if it asked for an unacceptable header.

For example, previously, sending 'Accept: bogus/bogus' would return 406,
but 'Accept: bogus' would crash the server. Now both return 406.