Remove CalendarLib dependency
For timestamp processing in HTTP headers so far the calendar module
has been used, which depends on unix.  This change implements a
simple RFC1123/RFC5322 parser and allows different clock sources to
be used.  Therefore a CLOCK module must be defined, that contains a
function "now" that returns the current time in seconds since epoch
as an int.  The Make functor now takes an additional Clock parameter
for this module.  Examples for POSIX or MirageOS clocks are provided.

Closes #73
Closes #79