Simplify http date parser according to RFC 7231
Before the date has been parsed according to RFC 1123.  However, as
pointed out by issue #91, RFC 7231 defines a subset of this format
for HTTP dates (IMF-fixdate).  This change simplifies the parser,
assuming the IMF-fixdate format according to RFC 7231.  In violence
to RFC 7231 the legacy RFC 850 and asctime() formats are
intentionally not handled, as before.

Closes #91