cleanup: refactor run_op and related functions.
Move the rd-management part into [DS.run], and reuse that to implement
functions in the logic that always set based on what the operation
returned, and then further modify it based on the result.

This allocates more, but this is the internet, so who really cares?

Also got rid of most of that grotesque type annotation on the run_op
method.